Создает точку входа и делит код
